requirements.txt
dvc
dvc[gdrive]
pandas
numpy
pyyaml
sklearn
matplotlib
seaborn
flask
gunicorn
scikit-learn
cassandra-driver